Move hit indexes to 32 bits (cms-patatrack#583)
Add a counter for forlorn doublets.

+++ CUDADataFormats/Track/interface/PixelTrackHeterogeneous.h
@@ -16,7 +16,7 @@ public:
   static constexpr int32_t stride() { return S; }
 
   using Quality = trackQuality::Quality;
-  using hindex_type = uint16_t;
+  using hindex_type = uint32_t;
   using HitContainer = cms::cuda::OneToManyAssoc<hindex_type, S, 5 * S>;
 
   // Always check quality is at least loose!
